+++ scripts/publish/workspace_versions.mjs
@@ -1,7 +1,5 @@
 /*
  * Update all package versions to prepare for publishing a new version
- * TODO: ideally update "create-jsenv" when any of the dependencies in
- * packages/create-jsenv/demo-/package.json is updated
  */
 
 import { updateWorkspaceVersions } from "@jsenv/package-workspace"
@@ -10,4 +8,12 @@ import { rootDirectoryUrl } from "@jsenv/core/jsenv.config.mjs"
 
 await updateWorkspaceVersions({
   directoryUrl: rootDirectoryUrl,
+  packagesRelations: {
+    "create-jsenv": [
+      "jsenv-demo-node-package",
+      "jsenv-demo-web",
+      "jsenv-demo-web-preact",
+      "jsenv-demo-web-react",
+    ],
+  },
 })